Weapon data has been added to serve as a database for main status and sub-status. Completed weapon data modification and UI work through upgrades and level-up logic implementation. Started working on UI related to weapon limit breaking _ 메인 스테이터스와 서브 스테이터스의 데이터 베이스 역할을 할, 무기 데이터 추가. 업그레이드를 통한 무기 데이터 수정 및 UI 작업과 레벨업 로직 구현 작업 완료. 무기 한계 돌파와 관련된 UI 작업 시작

+++ Assets/01Scripts/Item/WeaponAndEquipCls.cs
@@ -36,7 +36,7 @@ public class WeaponAndEquipCls : ItemClass
     public void SetMaxExp(int nMaxExp) { this.nMaxExp = nMaxExp; }
     public void SetLimitLevel(int nLimitLevel) { this.nLimitLevel = nLimitLevel;}
     public void SetEffectLevel(int nEffectLevel) { this.nEffectLevel = nEffectLevel; }
-    public void SetSubStat(int nSubStat) { this.nSubStat = nSubStat;}
+    public void SetSubStat(float nSubStat) { this.nSubStat = nSubStat;}
     public void SetMainStat(int nMainStat) { this.nMainStat = nMainStat; }
     public void SetEffectText(string sEffectText) { this.sEffectText = sEffectText; }
